ml/requirements.txt
pillow
numpy
requests
tqdm
pathlib
beautifulsoup4
aiohttp
